--- scripts/build-webpack-utils.js
@@ -1,2 +1,0 @@
-const tsTask = require('./tasks/ts');
-tsTask({ commonjsOnly: true });
