+++ scripts/just-tasks/rig.js
@@ -7,8 +7,8 @@ const { tslint } = require('./tslint');
 const { webpack } = require('./webpack');
 const { outdated, selfupdate } = require('./outdated');
 const { apiExtractor } = require('./apiExtractor');
-const buildCodepenExamples = require('../tasks/build-codepen-examples');
-const lintImports = require('../tasks/lint-imports');
+const buildCodepenExamples = require('./build-codepen-examples');
+const lintImports = require('./lint-imports');
 
 exports.rig = {
   clean,
