--- scripts/just-tasks/tslint.js
@@ -1,4 +1,0 @@
-// @ts-check
-
-const { tslintTask } = require('just-task-preset');
-exports.tslint = tslintTask();
