--- scripts/just-tasks/webpack.js
@@ -1,4 +1,0 @@
-// @ts-check
-
-const { webpackTask } = require('just-task-preset');
-exports.webpack = webpackTask();
