+++ scripts/tasks/apiExtractor.js
