+++ scripts/tasks/clean.js
