+++ scripts/tasks/jest.js
@@ -1,42 +1,11 @@
-module.exports = function(options) {
-  const path = require('path');
-  const fs = require('fs');
-  const exec = require('../exec-sync');
-  const findConfig = require('../find-config');
-  const jestConfigPath = findConfig('jest.config.js');
-  const resolve = require('resolve');
+// @ts-check
 
-  if (fs.existsSync(jestConfigPath)) {
-    const jestPath = resolve.sync('jest/bin/jest');
+const { argv } = require('just-task');
+const { jestTask } = require('just-task-preset');
 
-    const args = [
-      // Specify the config file.
-      `--config ${jestConfigPath}`,
-
-      // When there are no tests we still want to consider that a success.
-      // packages like `variants` do not have any tests (yet).
-      '--passWithNoTests',
-
-      // Forces test results output highlighting even if stdout is not a TTY.
-      '--colors',
-
-      // On Travis, run tests in serial as supposedly, the free Travis build terminates if multiple processes are spun up.
-      process.env.TRAVIS ? `--runInBand` : undefined,
-
-      // In production builds, produce coverage information.
-      (options.isProduction || process.env.TRAVIS) && '--coverage',
-
-      // If the -u flag is passed, pass it through.
-      options.argv && options.argv.indexOf('-u') >= 0 ? '-u' : '',
-
-      // Pass in custom arguments.
-      options.args
-    ]
-      .filter(arg => !!arg)
-      .join(' ');
-
-    const command = `node ${jestPath} ${args}`;
-
-    return exec(command, undefined, path.dirname(jestConfigPath), process);
-  }
-};
+exports.jest = () =>
+  jestTask({
+    ...(process.env.TRAVIS && { runInBand: true }),
+    ...(process.env.TRAVIS || argv().production ? { coverage: true } : undefined),
+    ...(argv().u || argv().updateSnapshot ? { updateSnapshot: true } : undefined)
+  });
